fixed typo

+++ phizzyForReal.py
@@ -13,7 +13,7 @@ parser.add_argument('--e', type=int, default=10000,
                     help='Number of epochs to train.')
 parser.add_argument('--d', type=int, default=100,
                     help='Number of data points generated for training.')
-parser.add_argument('--l', type=int, default=.001,
+parser.add_argument('--l', type=float, default=.001,
                     help='The learning rate.')
 args = parser.parse_args()
